buffer_cache: Refactor buffer manager with sparse arenas (#5047)
* buffer_cache: Refactor buffer manager with sparse arenas

The buffer cache is responsble for mirroring the GPU accessed address space in device local memory. On main it uses a simple scheme where the address space is divided into 16KB pages.
Buffers are allocated with that granularity and all pages of that buffer point to the same buffer object. When a range overlaps multiple buffers, a join overlap operation is performed
where a new buffer is created and the old buffer contents are copied over to the new one.

This scheme is quite simple and works well but has some issues

* It requires a 2-pass approach to buffer binding, because different bindings in the same draw/dispatch can cause a join overlap between them.
  If one binding was a storage buffer being written to, then data would be lost otherwise. This also causes various other issues when trying to batch memory uploads

* Join overlap can cause large VRAM usage spikes and requires heuristics to avoid repeated recreations.
  If the cache holds a large 1GB+ buffer, and then a range that partially overlaps it is accessed, the full 1GB range + the non overlapped size is going to get allocated.
  The old buffer will eventually be freed but this is still a large VRAM spike that can OOM or force system memory usage on lower VRAM gpus

To resolve these issues a new buffer management scheme is proposed in this pull request, that uses sparse residency buffers.
The address space is divided into 2^32 VA arenas, and each arena holds a certain amount of blocks/tiles which are determined by the sparse bind granularity of the host GPU.

The arena size has been chosen because some lower end apple sillicon GPUs only support creating buffers up to 4GB, even though the GPU hardware support much large address spaces.
Currently the code assumes that arena merges are always successfull, but that case will fail on those GPUs, for that bridge VA buffers need to be created and cached.
However accesses that straddle arenas are expected to be rare enough for this to not be an actual issue in practice.

Any buffer access is checked and overlapping blocks are made resident on demand by allocating memory. Residency requests are batched and submitted before scheduler submits the main command buffer.
If an access overlaps two adjacent arenas, then the arenas are merged into a single larger one and all their backings are brought up.

This new design avoids the old issues. There is no join operation aside from arena joining, which doesnt allocate any memory and is mostly for bookeeping purposes.
The buffer objects are also stable now so the 2-pass approach has been dropped.

For tracking residency a custom sorted vector based interval set has been added which is optimized for fast queries, as checking if a range is resident is in quite a hot path.

* vk_rasterizer: Move vertex and index buffer bind to rasterizer

Feels cleaner in general and wont have to pass barriers

* video_core: Add a new buffer barrier tracker

* buffer_cache: Fix arena migration bug

* video_core: Fully refactor resource management and barrier tracking

Introduces a new runtime class which is responsible for resource operations like buffer copy/fill, image upload/download/copy/resolve etc.
Having a centralized object makes the design significantly easier to maintain and extend. We dont have to pass scheduler in buffer/image objects anymore or blit helper in images.
Also barrier batching is now possible since we have a centralized place to track which buffer ranges were read/written. Image barriers can be batched as well before copies and also with buffer barriers.

* renderer_vulkan: Remove shader hle

New buffer changes make it mostly worthless aside from having vram build slower

* buffer_cache: Don't use VMA for residency allocations

We can avoid some cases of OOM where VMA would try to allocate more behind the scenes. But otherwise we can achieve a bit less memory usage with this

* video_core: Minor barrier fixes

* vk_rasterizer: Fix stupid ahh bug

* vk_runtime: Clear resources array

* buffer_cache: Fix partially null arena merges

* vk_rasterizer: Correct pipeline stage of accessed buffers

* renderer_vulkan: Add dxvk red-black tree for barrier tracking

Sorted vector interval sets are fast for queries but for barriers we have almost 50/50 split between query and insert traffic, so a more balanced structure is needed. the dxvk red-black tree seems pretty optimized for this use case so try it out

* texture_cache: Update size when resolving color depth alias

* renderer_vulkan: Add new staging pool to replace fixed buffers

Instead of allocating a large 512MB host allocation for uploads, now pool consists of 16MB blocks, any request below 16MB will use a ring of those blocks, otherwise it will allocate a dedicate large buffer that will be freed soon

* Revert "renderer_vulkan: Remove shader hle"

Will do it later

This reverts commit 52885faf3e87ba8cf19a29209ebb7164579d2e24.

* fix LICENSE

(cherry picked from commit 965c97c8272119d74251f98f76112b67686b5a2a)

Fork adaptation (cherry-picked onto feature/malos/sparse_arena):
- Buffers follow upstream: sparse arenas, Runtime barrier tracker, staging pool.
  Vertex, index and indirect reads are also registered with the tracker so a
  later write waits for them.
- Images keep the fork's Image operations (render scale, scaled blits, fused
  readback, MSAA backing): Runtime image calls delegate to them and record their
  own barriers; the buffers they touch are registered with the tracker.
- Pending sparse binds are queued from the submit callback before the recorder
  drain, under the queue mutex shared with the submission worker, and reported
  to the timeline completion (Android).
- A batched barrier under a hoisted operation ends the held pass first; the
  per-buffer barrier hoisting is removed.
- Copy-shader HLE keeps hoisting, guest mirroring and the deferred guest commit,
  now reading back through a deferred staging allocation. Staging releases are
  thread-safe (write-backs run on the priority thread).

+++ src/video_core/texture_cache/tile_manager.cpp
@@ -401,7 +401,7 @@ TileManager::Result TileManager::DetileImage(vk::Buffer in_buffer, u32 in_offset
 }
 
 void TileManager::TileImage(Image& in_image, std::span<vk::BufferImageCopy> buffer_copies,
-                            vk::Buffer out_buffer, u32 out_offset, u32 copy_size) {
+                            vk::Buffer out_buffer, u64 out_offset, u32 copy_size) {
     const auto& info = in_image.info;
     if (!info.props.is_tiled) {
         for (auto& copy : buffer_copies) {
@@ -527,7 +527,7 @@ vk::Pipeline TileManager::GetImageTilingPipeline(const ImageInfo& info, const Re
 }
 
 bool TileManager::TileImageFromScaled(Image& in_image, u32 num_mips, vk::Buffer out_buffer,
-                                      u32 out_offset) {
+                                      u64 out_offset) {
     const auto& info = in_image.info;
     if (!fused_readback || !in_image.IsScaled() || in_image.DroppedMips() ||
         in_image.IsReencoded() || !info.props.is_tiled || info.props.is_block ||
